requirements.txt
fastapi==0.86.0
kiteconnect==4.1.0
python-dotenv==0.21.0
selenium==4.2.0
pyotp
webdriver-manager==3.8.4
pandas
packaging==21.3
uvicorn
websockets